Successful refactor of existing functionality
