+++ src/dependencies/default_lib/default_funcs.h
@@ -1,0 +1,16 @@
+#ifndef DEFAULT_FUNCS_H
+#define DEFAULT_FUNCS_H
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <windows.h>
+
+/**
+ * Built-in command function declarations
+ */
+int spit_help(char **args);
+int spit_clear(char **args);
+int spit_exit(char **args);
+
+
+#endif // DEFAULT_FUNCS_H
