+++ src/dependencies/filetree_lib/filetree_funcs.h
@@ -1,0 +1,15 @@
+#ifndef FILETREE_FUNCS_H
+#define FILETREE_FUNCS_H
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <windows.h>
+
+/**
+ * Built-in command function declarations
+ */
+int spit_cd(char **args);
+int spit_ls(char **args);
+int spit_dir(char **args);
+
+#endif // FILETREE_FUNCS_H
